Cast size_t to int before negating it, to silence warnings
We can be pretty confident that the size of the two structs will be in
int's range, so the cast will preserve its value.

+++ src/conv-com/petable.C
@@ -653,7 +653,7 @@ PTvectorlist PeTable :: ExtractAndVectorize(comID id, int ufield, int npe, int *
   result->msgs  = (char**)((char*)result->sizes + (npacked+1)*sizeof(int) + sizeof(CmiChunkHeader));
 
   SIZEFIELD(result->sizes) = (npacked+1)*sizeof(int);
-  REFFIELD(result->sizes) = - (sizeof(struct ptvectorlist) + sizeof(CmiChunkHeader));
+  REFFIELD(result->sizes) = - (int)(sizeof(struct ptvectorlist) + sizeof(CmiChunkHeader));
   SIZEFIELD(result->msgs) = (npacked+1)*sizeof(int);
   REFFIELD(result->msgs) = - (sizeof(struct ptvectorlist) + (npacked+1)*sizeof(int) + 2*sizeof(CmiChunkHeader));
   CmiReference(result);
@@ -759,7 +759,7 @@ PTvectorlist PeTable :: ExtractAndVectorizeAll(comID id, int ufield) {
   result->msgs  = (char**)((char*)result->sizes + (nmsgs+1)*sizeof(int) + sizeof(CmiChunkHeader));
 
   SIZEFIELD(result->sizes) = (nmsgs+1)*sizeof(int);
-  REFFIELD(result->sizes) = - (sizeof(struct ptvectorlist) + sizeof(CmiChunkHeader));
+  REFFIELD(result->sizes) = - (int)(sizeof(struct ptvectorlist) + sizeof(CmiChunkHeader));
   SIZEFIELD(result->msgs) = (nmsgs+1)*sizeof(int);
   REFFIELD(result->msgs) = - (sizeof(struct ptvectorlist) + (nmsgs+1)*sizeof(int) + 2*sizeof(CmiChunkHeader));
   CmiReference(result);
